chore: wrap suspense of TesterLoginPage
